feat: implement not found page for invalid route requests
