--- src/app/provider/mantine/index.ts
@@ -1,1 +1,0 @@
-export { theme } from './mantine-theme';
