--- src/app/provider/mantine/mantine-theme.ts
@@ -1,5 +1,0 @@
-import { createTheme } from '@mantine/core';
-
-export const theme = createTheme({
-  primaryColor: 'dark',
-});
