+++ src/app/providers/mantine/index.ts
@@ -1,0 +1,1 @@
+export { theme } from './mantine-theme';
