+++ src/app/providers/mantine/mantine-theme.ts
@@ -1,0 +1,5 @@
+import { createTheme } from '@mantine/core';
+
+export const theme = createTheme({
+  primaryColor: 'dark',
+});
