+++ src/app/providers/router/routes.ts
@@ -1,8 +1,9 @@
 import AboutPage from '@pages/about/about';
 import BasketPage from '@pages/basket/basket';
 import CatalogPage from '@pages/catalog/catalog';
-import LoginPage from '@pages/login/login';
+import { LoginPage } from '@pages/login/index';
 import MainPage from '@pages/main/main';
+import { NotFoundPage } from '@pages/not-fount/index';
 import ProductPage from '@pages/product/product';
 import RegistrationPage from '@pages/registration/registration';
 import { Paths, RouteType } from '@shared/types/routerTypes';
@@ -13,6 +14,7 @@ export const mainRoutes: RouteType[] = [
   { path: Paths.About, element: AboutPage },
   { path: Paths.Basket, element: BasketPage },
   { path: Paths.Product, element: ProductPage },
+  { path: Paths.Error, element: NotFoundPage },
 ];
 
 export const authRoutes: RouteType[] = [
