+++ src/features/session/login/index.ts
@@ -1,1 +1,1 @@
-export { LoginForm } from './ui/login-form/login-form';
+export { LoginForm } from './ui/login-form';
