+++ src/features/session/login/ui/login-form.module.css
@@ -1,0 +1,6 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  align-items: center;
+}
