--- src/features/session/login/ui/login-form/login-form.module.css
@@ -1,6 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  align-items: center;
-}
