--- src/pages/login/login.module.css
@@ -1,1 +1,0 @@
-/* TODO: styles will be added later */
