--- src/pages/login/login.tsx
@@ -1,5 +1,0 @@
-const LoginPage = () => {
-  return <h2>Login Page</h2>;
-};
-
-export default LoginPage;
