+++ src/pages/login/ui/login.module.css
@@ -1,0 +1,19 @@
+.login {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+}
+
+.login-container {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  align-items: center;
+  justify-content: center;
+
+  padding: 30px 40px;
+  border-radius: 20px;
+
+  box-shadow: 0 4px 12px rgb(0 0 0 / 10%);
+}
