+++ src/pages/login/ui/login.tsx
@@ -1,0 +1,13 @@
+import { LoginForm } from '@features/session/login';
+import style from './login.module.css';
+
+export const LoginPage = () => {
+  return (
+    <div className={style.login}>
+      <div className={style['login-container']}>
+        <h1>Login</h1>
+        <LoginForm />
+      </div>
+    </div>
+  );
+};
