--- src/pages/login/ui/login/login.tsx
@@ -1,10 +1,0 @@
-import { LoginForm } from '../../../../features/session/login';
-
-export const Login = () => {
-  return (
-    <div>
-      <h1>Login</h1>
-      <LoginForm />
-    </div>
-  );
-};
