+++ src/pages/not-fount/index.ts
@@ -1,0 +1,1 @@
+export { NotFoundPage } from './ui/not-found-page';
