+++ src/pages/not-fount/ui/not-found-page.module.css
@@ -1,0 +1,56 @@
+.error-page {
+  display: grid;
+  grid-template-areas: 'content image';
+  grid-template-columns: 2fr 3fr;
+  gap: 20px;
+  align-items: center;
+  justify-content: center;
+
+  padding: 10px 20px;
+
+  text-align: center;
+}
+
+.error-page-content {
+  grid-area: content;
+}
+
+.error-page-image {
+  grid-area: image;
+}
+
+.error-page-image img {
+  max-width: 100%;
+  height: auto;
+}
+
+.button-link {
+  display: inline-block;
+
+  margin-top: 1rem;
+  padding: 0.75rem 1.5rem;
+  border-radius: 8px;
+
+  color: white;
+  text-decoration: none;
+
+  background-color: #1d3557;
+}
+
+.button-link:hover {
+  background-color: #457b9d;
+}
+
+@media (width <= 1000px) {
+  .error-page {
+    grid-template-areas:
+      'image'
+      'content';
+    grid-template-columns: 1fr;
+  }
+
+  .error-page-image {
+    width: 70%;
+    margin: 0 auto;
+  }
+}
